GSEIP-19 se guardan cambios

+++ controller/uploadsController.php
@@ -55,9 +55,11 @@ switch ($operation) {
             {
                 $temp = explode(".", $_FILES["myfile"]["name"]);
                 $newfilename = $id . '_' . $temp[0] . '_' .round(microtime(true)) . '.' . end($temp);
-                if(move_uploaded_file($_FILES["myfile"]["tmp_name"], $output_dir . $newfilename))
-                    RenovacionPersonal::uploadsUpload($output_dir, $newfilename, $id); //inserta en la BD
-                $ret[]= $newfilename;
+                if (!file_exists($output_dir . $newfilename)) {
+                    if(move_uploaded_file($_FILES["myfile"]["tmp_name"], $output_dir . $newfilename))
+                        RenovacionPersonal::uploadsUpload($output_dir, $newfilename, $id); //inserta en la BD
+                    $ret[]= $newfilename;
+                }
             }
             else  //Multiple files, file[]
             {
@@ -66,9 +68,11 @@ switch ($operation) {
                 {
                     $temp = explode(".", $_FILES["myfile"]["name"][$i]);
                     $newfilename = $id . '_' . $temp[0] . '_' .round(microtime(true)) . '.' . end($temp);
-                    if(move_uploaded_file($_FILES["myfile"]["tmp_name"][$i], $output_dir . $newfilename))
-                        RenovacionPersonal::uploadsUpload($output_dir, $newfilename, $id); //inserta en la BD
-                    $ret[]= $newfilename;
+                    if (!file_exists($output_dir . $newfilename)) {
+                        if (move_uploaded_file($_FILES["myfile"]["tmp_name"][$i], $output_dir . $newfilename))
+                            RenovacionPersonal::uploadsUpload($output_dir, $newfilename, $id); //inserta en la BD
+                        $ret[] = $newfilename;
+                    }
                 }
 
             }
